Separate to ParserUtils
