+++ src/Ipe/Parser/Import.hs
@@ -4,8 +4,8 @@ module Ipe.Parser.Import (singleParser, listParser) where
 
 import qualified Control.Monad
 import qualified Ipe.Grammar
-import Ipe.Parser (Parser)
-import qualified Ipe.Parser
+import Ipe.Parser.Utils (Parser)
+import qualified Ipe.Parser.Utils as Ipe.Parser
 import Text.Megaparsec ((<?>), (<|>))
 import qualified Text.Megaparsec as Parsec.Common
 import qualified Text.Megaparsec.Char as Parsec.Char
