+++ src/Ipe/Parser/TopLevelDefinition.hs
@@ -6,10 +6,10 @@ import qualified Control.Monad
 import Data.Text (Text)
 import qualified Data.Text as T
 import qualified Ipe.Grammar
-import Ipe.Parser (Parser)
-import qualified Ipe.Parser
 import qualified Ipe.Parser.Expression
 import qualified Ipe.Parser.TypeDefinition
+import Ipe.Parser.Utils (Parser)
+import qualified Ipe.Parser.Utils as Ipe.Parser
 import qualified Text.Megaparsec as Parsec.Common
 import qualified Text.Megaparsec.Char as Parsec.Char
