+++ src/Ipe/Parser/TypeDefinition.hs
@@ -6,8 +6,8 @@ import qualified Control.Monad
 import qualified Data.Map.Strict as Map
 import Data.Text (Text)
 import qualified Ipe.Grammar
-import Ipe.Parser (Parser)
-import qualified Ipe.Parser
+import Ipe.Parser.Utils (Parser)
+import qualified Ipe.Parser.Utils as Ipe.Parser
 import Text.Megaparsec ((<?>))
 import qualified Text.Megaparsec as Parsec.Common
 import qualified Text.Megaparsec.Char as Parsec.Char
